continued with itemComponents
